feat:Supports custom global setters, checks if arg type is supported, restructured file
__DESCRIPTION:__
Added support for giving the registerVariable function a custom setter. cons of it is that
it must be a global functions at the moment. No lambdas and no member functions.

All supported types are listed in the arguments.h file and the argument on construction checks
if what is being used is a supported type for parsing

Reordered part of the file for arguments for cleaner look